Add tests for thread names after attaching, currently failing
Right after attaching to a target that does not stop on attaching, the
threads request currently fails to return the correct thread names,
because they are not known to the adapter yet and are not retrieved from
the target until it stops.

The tests are only enabled under those circumstances where the problem
can easily be fixed: local non-stop, local all-stop async, remote
non-stop.

Signed-off-by: Christian Walther <walther@indel.ch>