Fix erroneous assumption of async mode when target does not support it
When async mode is requested using `-gdb-set mi-async on`, check with
GDB whether the request was actually granted once connected to a target.
It may not be, depending on the capabilities of the target (in
particular, the "native" target on Windows does not support async mode),
and in such a case, later decisions based on `gdb.getAsyncMode()` would
be wrong.

Signed-off-by: Christian Walther <walther@indel.ch>