Fix detach getting stuck on exited program (#437)
* Add test for detaching after running program exited, currently failing

When the debugged program has exited after it was, to the adapter's
knowledge, running, a `disconnect` request gets stuck and never
completes, because `GDBTargetDebugSession.doDisconnectRequest` tries to
pause the program before issuing the `disconnect` command to GDB, but
the stop notification that it expects as confirmation that the pause has
taken effect never arrives, because there are no threads anymore that
could stop.

This happens here in the afterEach function which first kills the
gdbserver, causing GDB to consider the program exited, and then in
`DebugClient.stop()` sends a disconnect request. It causes the test to
fail with a timeout in the "after each" hook.

The problem was not noticed in the existing tests because they all
happen to end with the program suspended, so no pausing is attempted.

Signed-off-by: Christian Walther <walther@indel.ch>

* Fix detach getting stuck on exited program

When attempting to pause a target that has no threads anymore because
the program has exited in the meantime, don't wait for a stop
notification that will never come.

There is still a race condition in that the threads could disappear
after we ask for them, but the likelihood is reduced.

This makes the test from the previous commit pass.

Signed-off-by: Christian Walther <walther@indel.ch>

* Remove the test timeout reduction hack from #407

The symptom it worked around was likely caused by the issue fixed in the
previous commit, so it is probably not needed anymore.

Signed-off-by: Christian Walther <walther@indel.ch>

---------

Signed-off-by: Christian Walther <walther@indel.ch>