Optionally add auxiliary GDB connection
This makes use of the capabilities of some
GDB remote servers that allow multiple connections to the same TCP/IP port and hence the same debug target. The second connection is in a paused state even if the main connection is running. Allowing to for example read memory and evaluate expressions in this state which would normally require non-stop mode support.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>