Fix missing thread names when attaching to a non-stopping target (#439)
* Add tests for thread names after attaching, currently failing

Right after attaching to a target that does not stop on attaching, the
threads request currently fails to return the correct thread names,
because they are not known to the adapter yet and are not retrieved from
the target until it stops.

The tests are only enabled under those circumstances where the problem
can easily be fixed: local non-stop, local all-stop async, remote
non-stop.

Signed-off-by: Christian Walther <walther@indel.ch>

* Fix missing thread names when attaching to a non-stopping target

Right after attaching to a target that does not stop on attaching, the
threads request would fail to return the correct thread names, because
they were not known to the adapter yet and were not retrieved from the
target until it stops. At that point, the adapter only knew about the
threads from their `thread-created` notifications, which do not include
the name.

Doing the required -thread-info command not only when the target is
stopped, but additionally under those circumstances where it is allowed
while the target is running (local non-stop, local all-stop async,
remote non-stop), fixes that under those circumstances.

Under the remaining circumstances (local all-stop sync, remote all-stop
async, remote all-stop sync) the behavior is unmodified, as explicitly
stopping the target every time to allow the -thread-info command is
probably undesired.

This makes the tests from the previous commit pass.

Signed-off-by: Christian Walther <walther@indel.ch>

---------

Signed-off-by: Christian Walther <walther@indel.ch>