Support auxiliary GDB connections to allow selected operations while CPU running (#449)
* Add named logger

Prepares for logging from multiple GDB backends
and to be able to distinguish between them.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Some refactoring for reuse with
auxiliary GDB connections.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Optionally add auxiliary GDB connection

This makes use of the capabilities of some
GDB remote servers that allow multiple connections to the same TCP/IP port and hence the same debug target. The second connection is in a paused state even if the main connection is running. Allowing to for example read memory and evaluate expressions in this state which would normally require non-stop mode support.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Route memoryRead and memoryWrite requests
through auxiliary GDB connection if available and
target is running.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Extend varManager to treat undefined frameRef as
global frame context.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Route evaluateRequest and variablesRequest through
auxiliary GDB if target is running. The latter to be able
to expand objects that are a result of evaluateRequest
while the target is running.

In this commit, evaluateRequest still blocks expression evaluation without frame reference. To be relaxed in next commit.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Allow expressions to go through aux GDB while
target is running. But stop CLI commands to pass
through. Commands that change the aux GDB state
can break the solution.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Fix sensitivity of expression evaluation
to leading/trailing spaces.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Relax blocking of GDB CLI commands in aux GDB mode,
GDB CLI commands are always routed through main GDB
connection.

Made conditions to deal with undefined frame info a little
more readable.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Auxiliary GDB tests, inspired by dynamic breakpoint options mock and test

* Adds auxiliary GDB mock adapter which creates
normal GDB backend for first connection and a mock
for the second. Implements mock functionality for relevant
sendCommand calls.
* Adds launch config to manually verify and learn which sendCommand variants need support in mock.
* Initial auxiliaryGdb test that does nothing but launch and end.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* aux tests: create client regardless of getting used

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Auxiliary GDB test implementations

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Ensure auxiliary GDB tests stop latest at main

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Validate applicability of auxiliaryGdb setting in combination
with gdbNonStop and gdbAsync. Includes tests.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Fix use of async in lateAsyncErrorsRemote test describe function

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Move launch request from beforeEach to test.
This allows to skip tests that would fail due to new
argument validation.

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

* Feedback jonahgraham

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>

---------

Signed-off-by: Jens Reinecke <jens.reinecke@arm.com>