Skip pausing target at startup if request has no breakpoints (#467)
In configuration phase, without this commit, breakpoint requests
are still sent to gdb even when there is no breakpoints to set.
When client does any of the various set breakpoint requests,
interrupt and continue command could be sent to gdb and
those commands could change the state of target unexpectedly.
Therefore this change makes it so breakpoint requests don't
pause the target during configuration phase if the request
has no breakpoints.

Co-authored-by: Dat Phan <dat.phan.xm@renesas.com>