Fix lint errors about "Forbidden non-null assertion"
regPC cannot be undefined at this point because we assert as much above, so
the non-null assertions are not needed. However the TypeScript compiler does
not know that and would error out because it does not understand
`expect(regPC).to.exist`. It does understand `assert(regPC !== undefined)`,
so replace it with that. (Not an exact replacement because `exist` also
checks for null, but that cannot occur here.)